requirements.txt
django>=5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
qrcode[pil]>=7.4
pillow>=10.0.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
whitenoise>=6.6.0
django-ratelimit>=4.1.0
shortuuid>=1.0.11
